Refactor KeyboardInputHandler and GamepadInputHandler
Create a common base class for both: do not make GamepadInputHandler
inherit from KeyboardInputHandler. This does not make sense.
